requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
easy-thumbnails==2.7.1
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
